--- transactions.txt
@@ -1,51 +1,0 @@
-2025-05-22 19:12:11 - Deposit - $500
-2025-05-22 19:12:11 - Withdraw - $150
-2025-05-22 19:12:11 - Deposit - $3002025-05-22 18:47:00 - Deposit - $500
-2025-05-22 18:47:00 - Withdraw - $150
-2025-05-22 18:47:00 - Deposit - $300
-2025-05-22 18:47:11 - Deposit - $500
-2025-05-22 18:47:11 - Withdraw - $150
-2025-05-22 18:47:11 - Deposit - $300
-2025-05-22 18:48:27 - Deposit - $500
-2025-05-22 18:48:27 - Withdraw - $150
-2025-05-22 18:48:27 - Deposit - $300
-2025-05-22 19:58:39 - Deposit - $500
-2025-05-22 19:58:39 - Deposit - $500
-2025-05-22 19:58:39 - Deposit - $500
-2025-05-22 19:58:39 - Deposit - $500
-2025-05-22 19:58:39 - Withdraw - $150
-2025-05-22 19:58:39 - Deposit - $500
-2025-05-22 19:58:39 - Withdraw - $150
-2025-05-22 19:58:39 - Withdraw - $150
-2025-05-22 19:58:39 - Withdraw - $150
-2025-05-22 19:58:39 - Withdraw - $150
-2025-05-22 19:58:39 - Deposit - $300
-2025-05-22 19:58:39 - Deposit - $300
-2025-05-22 19:58:39 - Deposit - $300
-2025-05-22 19:58:39 - Deposit - $300
-2025-05-22 19:58:39 - Deposit - $300
-2025-05-22 19:58:39 - Deposit - $300
-2025-05-22 19:58:49 - Deposit - $500
-2025-05-22 19:58:49 - Withdraw - $150
-2025-05-22 19:58:49 - Deposit - $300
-2025-05-22 19:58:54 - Deposit - $500
-2025-05-22 19:58:54 - Withdraw - $150
-2025-05-22 19:58:54 - Deposit - $300
-2025-05-22 19:58:55 - Deposit - $500
-2025-05-22 19:58:55 - Withdraw - $150
-2025-05-22 19:58:55 - Deposit - $300
-2025-05-22 19:59:00 - Deposit - $500
-2025-05-22 19:59:00 - Withdraw - $150
-2025-05-22 19:59:00 - Deposit - $300
-2025-05-22 20:54:43 - Deposit - $500
-2025-05-22 20:54:43 - Withdraw - $150
-2025-05-22 20:54:43 - Deposit - $300
-2025-05-22 20:55:12 - Deposit - $500
-2025-05-22 20:55:12 - Withdraw - $150
-2025-05-22 20:55:12 - Deposit - $300
-2025-05-22 20:56:24 - Deposit - $500
-2025-05-22 20:56:24 - Withdraw - $150
-2025-05-22 20:56:24 - Deposit - $300
-2025-05-22 21:03:11 - Deposit - $500
-2025-05-22 21:03:11 - Withdraw - $150
-2025-05-22 21:03:11 - Deposit - $300
